% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elec_prices.R
\docType{data}
\name{elec_prices}
\alias{elec_prices}
\title{Daily electricity price profiles from the Day-Ahead Spanish Electricity Market}
\format{A data frame with 365 rows and 24 variables:
\describe{
  \item{H1}{Electricity price for hour 1}
  \item{H2}{Electricity price for hour 2}
  ...
}}
\source{
\url{https://www.esios.ree.es/es/analisis/600}
}
\usage{
elec_prices
}
\description{
A dataset containing the hourly electricity prices for Spain
in the Day-Ahead Market (MIBEL)
}
\keyword{datasets}
