% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdaACF.R
\docType{package}
\name{fdaACF}
\alias{fdaACF}
\alias{fdaACF-package}
\title{fdaACF: Autocorrelation function for Functional Time Series}
\description{
The fdaACF package provides diagnostic and analysis tools
to quantify the serial autocorrelation across lags of a
given functional time series in order to improve the identification
and diagnosis of functional ARIMA models. The autocorrelation function
is based on the L2 norm of the lagged covariance operators of the series.
Several real-world datasets are included to illustrate the application
of these techniques.
}
